Move combine join to equivalences.rs file
